refactor: refactor dialog naming
